+++ src/main/java/org/compiere/orm/X_AD_IndexColumn.java
@@ -54,7 +54,7 @@ public class X_AD_IndexColumn extends PO implements IndexColumn {
      * @return Column in the table
      */
     public int getColumnId() {
-        Integer ii = (Integer) getValue(COLUMNNAME_AD_Column_ID);
+        Integer ii = getValue(COLUMNNAME_AD_Column_ID);
         if (ii == null) return 0;
         return ii;
     }
