+++ src/main/java/org/compiere/orm/X_AD_User_Roles.java
@@ -51,7 +51,7 @@ public class X_AD_User_Roles extends PO implements UserRoles {
      * @return Responsibility Role
      */
     public int getRoleId() {
-        Integer ii = (Integer) getValue(COLUMNNAME_AD_Role_ID);
+        Integer ii = getValue(COLUMNNAME_AD_Role_ID);
         if (ii == null) return 0;
         return ii;
     }
@@ -72,7 +72,7 @@ public class X_AD_User_Roles extends PO implements UserRoles {
      * @return User within the system - Internal or Business Partner Contact
      */
     public int getUserId() {
-        Integer ii = (Integer) getValue(COLUMNNAME_AD_User_ID);
+        Integer ii = getValue(COLUMNNAME_AD_User_ID);
         if (ii == null) return 0;
         return ii;
     }
